src/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
authlib
python-jose[cryptography]
psycopg2-binary
